backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (PostgreSQL) - Keep for production option
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# AI & External APIs
groq==0.4.1
httpx==0.25.2

# Core Python packages
pydantic[email]==2.5.0
email-validator==2.1.0

# Utilities
aiofiles==23.2.1
jinja2==3.1.2
Pillow==10.1.0
python-decouple==3.8
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1